✓ adds cast refs #31
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui@1885 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ WORK_01/common/arcunzip.c
@@ -194,7 +194,7 @@ static long method0_seek(ARCFH arcfh, long pos, UINT method) {
 	else if (pos > (long)m0->size) {
 		pos = m0->size;
 	}
-	m0->pos = pos;
+	m0->pos = (UINT)pos;
 	return(pos);
 }
 
@@ -242,7 +242,7 @@ typedef struct {
 	UINT		srcpos;
 	UINT		dstsize;
 	UINT		dstpos;
-	long		dstdiff;
+	int			dstdiff;
 
 	z_stream	stream;
 	int			err;
@@ -363,7 +363,7 @@ static long method8_seek(ARCFH arcfh, long pos, UINT method) {
 	else if (pos > (long)m8->dstsize) {
 		pos = m8->dstsize;
 	}
-	m8->dstdiff = pos - m8->dstpos;
+	m8->dstdiff = (int)(pos - m8->dstpos);
 	return(pos);
 }
